Fix cache issues and refactor code. (594)
