+++ lira/api/notifications.py
@@ -41,7 +41,6 @@ def post(body):
     cromwell_submission = current_app.prepare_submission(wdl_config, lira_config.submit_wdl)
     logger.info(current_app.prepare_submission.cache_info())
 
-
     dry_run = getattr(lira_config, 'dry_run', False)
     if dry_run:
         logger.warning('Not launching workflow because Lira is in dry_run mode')
@@ -53,7 +52,7 @@ def post(body):
     else:
         cromwell_response = cromwell_tools.start_workflow(
             wdl_file=cromwell_submission.wdl_file,
-            zip_file=cromwell_submission.wdl_deps_file,
+            zip_file=cromwell_tools.make_zip_in_memory(cromwell_submission.wdl_deps_dict),
             inputs_file=cromwell_inputs_file,
             inputs_file2=cromwell_submission.wdl_static_inputs_file,
             options_file=cromwell_submission.options_file,
@@ -71,55 +70,3 @@ def post(body):
             status_code = 201
 
     return lira_utils.response_with_server_header(response_json, status_code)
-
-
-def create_prepare_submission_function(cache_wdls):
-    """Returns decorated prepare_submission function. Decorator is determined as follows:
-    Python 2: Always decorate with noop_lru_cache, since functools.lru_cache is not available in 2.
-    Python 3: Use functools.lru_cache if cache_wdls is true, otherwise use noop_lru_cache."""
-
-    # Use noop_lru_cache unless cache_wdls is true and functools.lru_cache is available
-    lru_cache = lira_utils.noop_lru_cache
-    if not cache_wdls:
-        logger.info('Not caching wdls because Lira is configured with cache_wdls false')
-    else:
-        try:
-            from functools import lru_cache
-        except ImportError:
-            logger.info('Not caching wdls because functools.lru_cache is not available in Python 2.')
-
-    @lru_cache(maxsize=None)
-    # Setting maxsize to None here means each unique call to prepare_submission (defined by parameters used)
-    # will be added to the cache without evicting any other items. Additional non-unique calls
-    # (parameters identical to a previous call) will read from the cache instead of actually
-    # calling this function. This means that the function will be called no more than once for
-    # each wdl config, but we can have arbitrarily many wdl configs without running out of space
-    # in the cache.
-    def prepare_submission(wdl_config, submit_wdl):
-        """Load into memory all static data needed for submitting a workflow to Cromwell"""
-
-        # Read files into memory
-        wdl_file = cromwell_tools.download(wdl_config.wdl_link)
-        wdl_static_inputs_file = cromwell_tools.download(wdl_config.wdl_static_inputs_link)
-        options_file = cromwell_tools.download(wdl_config.options_link)
-
-        # Create zip of analysis and submit wdls
-        url_to_contents = cromwell_tools.download_to_map(wdl_config.analysis_wdls + [submit_wdl])
-        wdl_deps_file = cromwell_tools.make_zip_in_memory(url_to_contents)
-
-        return CromwellSubmission(wdl_file, wdl_static_inputs_file, options_file, wdl_deps_file)
-
-    return prepare_submission
-
-
-class CromwellSubmission(object):
-    """Holds static data needed for submitting a workflow to Cromwell, including
-    the top level wdl file, the static inputs json file, the options file,
-    and the dependencies zip file.
-    """
-
-    def __init__(self, wdl_file, wdl_static_inputs_file=None, options_file=None, wdl_deps_file=None):
-        self.wdl_file = wdl_file
-        self.wdl_static_inputs_file = wdl_static_inputs_file
-        self.options_file = options_file
-        self.wdl_deps_file = wdl_deps_file
