+++ lira/lira_utils.py
@@ -2,7 +2,10 @@
 """
 import json
 import logging
+from cromwell_tools import cromwell_tools
 from flask import make_response
+
+logger = logging.getLogger('lira.{module_path}'.format(module_path=__name__))
 
 
 def response_with_server_header(body, status):
@@ -79,3 +82,55 @@ def noop_lru_cache(maxsize=None, typed=False):
         wrapper.cache_clear = cache_clear
         return wrapper
     return real_noop_lru_cache
+
+
+def create_prepare_submission_function(cache_wdls):
+    """Returns decorated prepare_submission function. Decorator is determined as follows:
+    Python 2: Always decorate with noop_lru_cache, since functools.lru_cache is not available in 2.
+    Python 3: Use functools.lru_cache if cache_wdls is true, otherwise use noop_lru_cache."""
+
+    # Use noop_lru_cache unless cache_wdls is true and functools.lru_cache is available
+    lru_cache = noop_lru_cache
+    if not cache_wdls:
+        logger.info('Not caching wdls because Lira is configured with cache_wdls false')
+    else:
+        try:
+            from functools import lru_cache
+        except ImportError:
+            logger.info('Not caching wdls because functools.lru_cache is not available in Python 2.')
+
+    @lru_cache(maxsize=None)
+    # Setting maxsize to None here means each unique call to prepare_submission (defined by parameters used)
+    # will be added to the cache without evicting any other items. Additional non-unique calls
+    # (parameters identical to a previous call) will read from the cache instead of actually
+    # calling this function. This means that the function will be called no more than once for
+    # each wdl config, but we can have arbitrarily many wdl configs without running out of space
+    # in the cache.
+    def prepare_submission(wdl_config, submit_wdl):
+        """Load into memory all static data needed for submitting a workflow to Cromwell"""
+
+        # Read files into memory
+        wdl_file = cromwell_tools.download(wdl_config.wdl_link)
+        wdl_static_inputs_file = cromwell_tools.download(wdl_config.wdl_static_inputs_link)
+        options_file = cromwell_tools.download(wdl_config.options_link)
+
+        # Create zip of analysis and submit wdls
+        url_to_contents = cromwell_tools.download_to_map(wdl_config.analysis_wdls + [submit_wdl])
+        wdl_deps_dict = url_to_contents
+
+        return CromwellSubmission(wdl_file, wdl_static_inputs_file, options_file, wdl_deps_dict)
+
+    return prepare_submission
+
+
+class CromwellSubmission(object):
+    """Holds static data needed for submitting a workflow to Cromwell, including
+    the top level wdl file, the static inputs json file, the options file,
+    and the dependencies zip file.
+    """
+
+    def __init__(self, wdl_file, wdl_static_inputs_file=None, options_file=None, wdl_deps_dict=None):
+        self.wdl_file = wdl_file
+        self.wdl_static_inputs_file = wdl_static_inputs_file
+        self.options_file = options_file
+        self.wdl_deps_dict = wdl_deps_dict
